+++ pom.xml
@@ -21,8 +21,8 @@
     <artifactId>booster-parent</artifactId>
     <version>23</version>
   </parent>
-  <groupId>io.openshift.booster</groupId>
-  <artifactId>spring-boot-istio-security-parent</artifactId>
+  <groupId>dev.snowdrop.example</groupId>
+  <artifactId>istio-security</artifactId>
   <packaging>pom</packaging>
   <version>2.1.3-3-SNAPSHOT</version>
   <properties>
