+++ spring-boot-istio-security-greeting/pom.xml
@@ -20,8 +20,8 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.openshift.booster</groupId>
-    <artifactId>spring-boot-istio-security-parent</artifactId>
+    <groupId>dev.snowdrop.example</groupId>
+    <artifactId>istio-security</artifactId>
     <version>2.1.3-3-SNAPSHOT</version>
   </parent>
