+++ spring-boot-istio-security-greeting/src/main/java/dev/snowdrop/example/GreetingApplication.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster;
+package dev.snowdrop.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
