+++ spring-boot-istio-security-greeting/src/main/resources/static/index.html
@@ -48,7 +48,7 @@
             <div class="col-md-6">
                 <h3>Instructions</h3>
                 <p>
-                   Please reference the mission documentation or <a href="https://github.com/snowdrop/spring-boot-istio-security-booster/blob/master/README.adoc">README</a> for next steps and instructions to run each scenario.
+                   Please reference the mission documentation or <a href="https://github.com/snowdrop/istio-security-example/blob/master/README.adoc">README</a> for next steps and instructions to run each scenario.
                 </p>
                 <ul>
                    <li>Scenario #1. Mutual TLS</li>
