+++ spring-boot-istio-security-greeting/src/test/java/dev/snowdrop/example/service/GreetingControllerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 import org.junit.Before;
 import org.junit.Test;
