+++ spring-boot-istio-security-name/.openshiftio/application.yaml
@@ -55,7 +55,7 @@ objects:
       app: spring-boot-istio-security-name
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: runtime-spring-boot-istio-security-name
   spec:
     tags:
@@ -70,7 +70,7 @@ objects:
       app: spring-boot-istio-security-name
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-istio-security-name
   spec:
     output:
@@ -116,7 +116,7 @@ objects:
       app: spring-boot-istio-security-name
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-istio-security-name
   spec:
     ports:
@@ -127,7 +127,7 @@ objects:
     selector:
       app: spring-boot-istio-security-name
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
@@ -135,7 +135,7 @@ objects:
       app: spring-boot-istio-security-name
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-istio-security-name
   spec:
     replicas: 1
@@ -143,7 +143,7 @@ objects:
     selector:
       app: spring-boot-istio-security-name
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -156,7 +156,7 @@ objects:
           app: spring-boot-istio-security-name
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -214,7 +214,7 @@ objects:
       app: spring-boot-istio-security-name
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-istio-security-name
     annotations:
       kubernetes.io/ingress.class: istio
