+++ tests/pom.xml
@@ -3,8 +3,8 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>spring-boot-istio-security-parent</artifactId>
-    <groupId>io.openshift.booster</groupId>
+    <artifactId>istio-security</artifactId>
+    <groupId>dev.snowdrop.example</groupId>
     <version>2.1.3-3-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
@@ -96,12 +96,12 @@
             -->
             <dependencies>
               <dependency>
-                <groupId>io.openshift.booster</groupId>
+                <groupId>dev.snowdrop.example</groupId>
                 <artifactId>spring-boot-istio-security-greeting</artifactId>
                 <version>${project.version}</version>
               </dependency>
               <dependency>
-                <groupId>io.openshift.booster</groupId>
+                <groupId>dev.snowdrop.example</groupId>
                 <artifactId>spring-boot-istio-security-name</artifactId>
                 <version>${project.version}</version>
               </dependency>
